+++ src/shim/host/platform_host.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (C) 2025, Advanced Micro Devices, Inc. All rights reserved.
 
-#include "shim_debug.h"
+#include "../shim_debug.h"
 #include "platform_host.h"
 #include "core/common/trace.h"
 #include <fstream>
